refactor(groupUser): give invite links their own service module
The eight invite-link functions and the open-invite count move out of
groupUserServices into inviteLinkServices, in the same folder and sharing
its types. Invite link is already its own term in the glossary with its own
lifecycle; the filesystem now agrees.

The module stays inside groupUser rather than becoming a peer domain: an
invite exists only as a path into a group's membership, and promoting it
would mean relocating the invite notifier and splitting types the
open-invite query joins against.
